% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{center}
\usage{
center(x)
}
\arguments{
\item{x}{a vector of numeric or integer values}
}
\value{
a vector of numeric values
}
\description{
\code{center} returns a centered copy of a vector of values. To do this, it
subtracts the mean of the vector from each value.
}
\examples{
a <- c(5, 0, 5, 1, 2, 3, 0, 2, 2, 6)
center(a)
}
