% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\title{scale}
\usage{
scale(x)
}
\arguments{
\item{x}{a vector of numeric or integer values}
}
\value{
a vector of numeric values
}
\description{
\code{scale} returns a scaled version of a vector. To do this, it divides each
value of the vector by the standard deviation of the vector.
}
\examples{
a <- c(5, 0, 5, 1, 2, 3, 0, 2, 2, 6)
scale(a)
}
